requirements.txt
# Azure SDK and AI Services
azure-identity                                # Azure authentication and credentials management
azure-ai-projects==1.0.0b12                         # Azure AI project management
azure-ai-inference[opentelemetry]             # Azure AI inference with OpenTelemetry support
azure-ai-inference[prompts]
azure-ai-evaluation                           # Azure AI model evaluation tools
azure-ai-contentsafety                        # Azure content safety services
azure-search-documents==11.6.0b12                     # Azure AI Search integration
azure-ai-ml
azure-ai-agents==1.0.2


# Telemetry and Monitoring
azure-monitor-opentelemetry                  # Azure Monitor OpenTelemetry integration
azure-core-tracing-opentelemetry             # Azure core OpenTelemetry tracing
opentelemetry-exporter-otlp                  # OpenTelemetry protocol exporter
opentelemetry-sdk                            # OpenTelemetry SDK
opentelemetry-instrumentation-openai-v2 
opentelemetry-exporter-otlp-proto-grpc


# Jupyter and Development Tools
jupyter                                      # Jupyter notebook interface
python-dotenv                                # Environment variable management
notebook                                     # Jupyter notebook core
ipykernel                                    # IPython kernel for Jupyter
pandas                                       # Data manipulation library
tabulate                                     # Pretty-print tabular data
nbformat                                     # Notebook format conversion
asyncio                                      # Async I/O support
typing                                       # Type hinting support
aiofiles

# MCP protocol
mcp

# Black
black[jupyter]

# Semantic Kernel
aiortc
semantic-kernel==1.23.0


# Tracing
opentelemetry-sdk
opentelemetry-exporter-otlp